users消す前
