+++ app/helpers/my/users_helper.rb
@@ -1,0 +1,2 @@
+module My::UsersHelper
+end
