+++ app/views/layouts/application.html.erb
@@ -13,7 +13,7 @@
   <%= link_to "記事ランキングへ", articles_path %>
   <% if user_signed_in? %>
     <%= link_to "ログアウト", destroy_user_session_path, method: :delete %>
-    <%= link_to "マイページ", user_path(current_user) %>
+    <%= link_to "マイページ", my_user_path(current_user) %>
   <% else %>
     <%= link_to "新規登録", new_user_registration_path %>
     <%= link_to "ログイン", new_user_session_path %>
