+++ app/views/my/users/show.html.erb
@@ -1,7 +1,7 @@
 <h1>Users#show</h1>
 <%= @show_user.name %>
 <%= @show_user.email %>
-<%= form_with(model: @user,local: true) do |form| %>
+<%= form_with(model: [:my, @user],local: true) do |form| %>
   <% if @user.errors.any? %>
     <div id="error_explanation">
       <h2><%= pluralize(@user.errors.count, "error") %> prohibited this comment from being saved:</h2>
