+++ test/controllers/my/users_controller_test.rb
@@ -1,0 +1,7 @@
+require 'test_helper'
+
+class My::UsersControllerTest < ActionDispatch::IntegrationTest
+  # test "the truth" do
+  #   assert true
+  # end
+end
